requirements.txt
flask==2.0.1
opencv-python-headless==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
torch==1.9.0
transformers==4.11.3
easyocr==1.4.1
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.26.0
werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1 